Move edittable form for task items to be with rest of task forms and update references
