+++ app/manage/list/components/AddForm.tsx
@@ -4,11 +4,10 @@ import { FormSubmit } from "@/app/manage/list/components/HandleSubmitCRUD";
 import Link from "next/link";
 import useTaskListContext from "@/hooks/useTaskListContext";
 import { Dispatch, SetStateAction } from "react";
-import TaskItemEdittableFormLayout from "./TaskItemEdittableFormLayout";
+import TaskItemEdittableFormLayout from "../../../../components/TaskItemEdittableFormLayout";
 import { useQueryClient } from "react-query";
 
 import { useUserInfo } from "@/hooks/useUserInfo";
-
 
 interface AddFormProps {
   onBack: Dispatch<SetStateAction<boolean>>;
@@ -23,9 +22,7 @@ const AddForm: React.FC<AddFormProps> = ({ onBack }) => {
   const HandleSubmit: React.FormEventHandler<HTMLFormElement> = async (
     e: React.FormEvent<HTMLFormElement>
   ) => {
-
     FormSubmit(e, "add", user?.id);
-
 
     queryClient.invalidateQueries({ queryKey: ["Tasks"] });
