+++ app/manage/list/components/EditForm.tsx
@@ -5,7 +5,7 @@ import { Dispatch, SetStateAction } from "react";
 import Link from "next/link";
 import useTaskListContext from "@/hooks/useTaskListContext";
 import { Task } from "@/types/Task";
-import TaskItemEdittableFormLayout from "./TaskItemEdittableFormLayout";
+import TaskItemEdittableFormLayout from "../../../../components/TaskItemEdittableFormLayout";
 import { useQueryClient } from "react-query";
 
 interface EditFormProps {
